LPS-55772 Both Auth beans wired, only one createInstance will be invoked
https://issues.liferay.com/browse/LPS-55772

Conflicts:
	shared/solr4-shared/src/OSGI-INF/blueprint/solr-blueprint.xml --> We don't need this for 62x.

Author:    Tibor Lipusz <tibor.lipusz@liferay.com>
