+++ webs/solr4-web/docroot/WEB-INF/src/META-INF/solr-spring.xml
@@ -36,7 +36,7 @@
 
 	<!-- Solr search engine client -->
 
-	<bean id="com.liferay.portal.search.solr.http.HttpClientFactory" class="com.liferay.portal.search.solr.http.BasicAuthPoolingDelegatingHttpClientFactory" destroy-method="shutdown">
+	<bean id="com.liferay.portal.search.solr.http.HttpClientFactory" class="com.liferay.portal.search.solr.http.BasicAuthPoolingHttpClientFactory" destroy-method="shutdown">
 		<property name="defaultMaxConnectionsPerRoute" value="20" />
 		<property name="httpRequestInterceptors">
 			<list>
@@ -46,6 +46,30 @@
 		<property name="maxTotalConnections" value="20" />
 		<property name="password" value="${basic.auth.password}" />
 		<property name="username" value="${basic.auth.username}" />
+	</bean>
+	<bean class="com.liferay.portal.search.solr.http.CertAuthPoolingHttpClientFactory" id="com.liferay.portal.search.solr.http.CertAuthPoolingHttpClientFactory">
+		<property name="defaultMaxConnectionsPerRoute" value="20" />
+		<property name="httpRequestInterceptors">
+			<list>
+				<bean class="com.liferay.portal.search.solr.interceptor.PreemptiveAuthInterceptor" />
+			</list>
+		</property>
+		<property name="maxTotalConnections" value="20" />
+		<property name="sslSocketFactoryBuilder">
+			<bean class="com.liferay.portal.search.solr.http.SSLSocketFactoryBuilderImpl">
+				<property name="keyStoreLoader">
+					<bean class="com.liferay.portal.search.solr.http.KeyStoreLoaderImpl" />
+				</property>
+				<property name="keyStorePassword" value="${digital.signature.key.store.password}" />
+				<property name="keyStorePath" value="${digital.signature.key.store.path}" />
+				<property name="keyStoreType" value="${digital.signature.key.store.type}" />
+				<property name="trustStorePassword" value="${digital.signature.trust.store.password}" />
+				<property name="trustStorePath" value="${digital.signature.trust.store.path}" />
+				<property name="trustStoreType" value="${digital.signature.trust.store.type}" />
+				<property name="verifyServerCertificate" value="true" />
+				<property name="verifyServerHostname" value="true" />
+			</bean>
+		</property>
 	</bean>
 	<bean id="com.liferay.portal.search.solr.server.HttpSolrServer" class="com.liferay.portal.search.solr.server.HttpSolrServer">
 		<property name="httpClientFactory" ref="com.liferay.portal.search.solr.http.HttpClientFactory" />
